Includes new cocoon code interface that describes how a cocoon code should be constructed.
Cocoon codes must include `Init` and `Invoke` methods.

When the `Init` method is run and it returns an error, the cocoon code is immediately stopped.

The `Invoke` method is called whenever an invoke transaction is received. It can return result and an error. Result is coerced and sent to the requesting service.