Implemented cocoon code invocation via HTTP REST API.
Supports structured and unstructured invoke requests. A structured invoke request is one were the request content type is `application/json` and as a `function` value set. If a function value is not set, it is considered to be an unstructured invocation.

A unstructured invocation is a request whose body isn't JSON or its JSON body does not contain a `function` value.